Reduce location precision to float32
This is the practical limit of commercial data and should be more than enough for our use case while ideally providing better index performance.

+++ internal/entity/location.go
@@ -23,10 +23,10 @@ type Location struct {
 }
 
 // NewLocation creates a location using a token extracted from coordinate
-func NewLocation(lat, lng float64) *Location {
+func NewLocation(lat, lng float32) *Location {
 	result := &Location{}
 
-	result.ID = s2.Token(lat, lng)
+	result.ID = s2.Token(float64(lat), float64(lng))
 
 	return result
 }
